+++ CDJSVis/filtering/voro_iface.h
@@ -17,4 +17,4 @@ typedef struct {
 #ifdef __cplusplus
 extern "C"
 #endif
-int computeVoronoiVoroPlusPlusWrapper(int, double*, int*, double*, double*, int, double*, vorores_t*);
+int computeVoronoiVoroPlusPlusWrapper(int, double*, int*, double*, double*, int, double*, double, vorores_t*);